refactor(core): relocate formatEntityName out of hooks/ into string-utils
It's a plain function, not a hook -- the hooks/use-entity-name directory
was leftover naming from before the DisplayProvider removal. string-utils
already holds the other casing/formatting utilities (sentenceCaseEnumValue,
capitalizeFirstLetter), so it belongs there.
