+++ javascript/packages/core/utils/__tests__/string-utils.tests.ts
@@ -1,5 +1,6 @@
 import {
   capitalizeFirstLetter,
+  formatEntityName,
   isAbsoluteUrl,
   isNavigableUrl,
   safeStringify,
@@ -239,5 +240,35 @@ describe('safeStringify', () => {
       const deep = { level1: { level2: { level3: { value: 'deep' } } } };
       expect(safeStringify(deep)).toBe('{"level1":{"level2":{"level3":{"value":"deep"}}}}');
     });
+  });
+});
+
+describe('formatEntityName', () => {
+  it('title-cases for "nav"', () => {
+    expect(formatEntityName('trained models', 'nav')).toBe('Trained Models');
+  });
+
+  it('passes through for undefined casing', () => {
+    expect(formatEntityName('trained models', undefined)).toBe('trained models');
+  });
+
+  it('passes through when casing is omitted entirely', () => {
+    expect(formatEntityName('trained models')).toBe('trained models');
+  });
+
+  it('title-cases a single word', () => {
+    expect(formatEntityName('pipelines', 'nav')).toBe('Pipelines');
+  });
+
+  it('returns an empty string unchanged', () => {
+    expect(formatEntityName('', 'nav')).toBe('');
+  });
+
+  it('preserves acronyms already capitalized in the canonical name', () => {
+    expect(formatEntityName('AI agents', 'nav')).toBe('AI Agents');
+  });
+
+  it('preserves hyphens rather than splitting them into separate words', () => {
+    expect(formatEntityName('one-off predictions', 'nav')).toBe('One-off Predictions');
   });
 });
